Current game score successfully exported to game.js
