+++ lib/boat.js
@@ -1,5 +1,4 @@
 var ImageRenderer = require('./image-renderer');
-
 var imageRenderer = new ImageRenderer();
 var images = imageRenderer.init();
 
@@ -10,6 +9,7 @@ function Boat(options) {
   this.width = options.width || 10;
   this.height = options.height || 10;
   this.velocity = options.velocity || 5;
+  this.score = 0;
 }
 
 Boat.prototype.draw = function(context) {
@@ -21,7 +21,7 @@ Boat.prototype.moveBoatLeft = function(boat) {
   if (boat.x > 15) {
     boat.x -= boat.velocity;
     return boat;
-  } 
+  }
 };
 
 Boat.prototype.moveBoatRight = function(boat, canvasWidth) {
