+++ lib/bullet.js
@@ -1,5 +1,4 @@
 var ImageRenderer = require('./image-renderer');
-
 var imageRenderer = new ImageRenderer();
 var images = imageRenderer.init();
