+++ lib/collision.js
@@ -1,7 +1,6 @@
 function collision(fish, bullet) {
   if (xCollision(fish, bullet) && yCollision(fish, bullet)) {
     console.log("COLLISION");
-    updateScore(fish.y);
     console.log("fish y = " + fish.y)
     return true;
   }
@@ -23,14 +22,4 @@ function yCollision(fish, bullet) {
   }
 }
 
-var score = 0;
-var updateScore = function(value) {
-  console.log("beginning score = " + score);
-  return score += value;
-  // console.log("updated score = " + score);
-};
-
-
-module.exports.collision = collision;
-module.exports.score = score;
-module.exports.updateScore = updateScore;
+module.exports = collision
