+++ lib/game.js
@@ -1,34 +1,25 @@
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
-
-var canvasWidth =  canvas.width;
-var canvasHeight =  canvas.height;
+var canvasWidth = canvas.width;
+var canvasHeight = canvas.height;
 
 var ImageRenderer = require('./image-renderer');
-
 var imageRenderer = new ImageRenderer();
 var images = imageRenderer.init();
 
-var score = 0;
+var Boat = require('./boat');
+var Bullet = require('./bullet');
+var Fish = require('./fish').Fish;
+var fishies = require('./fish').fishies;
+var collision = require('./collision');
+var bullets = [];
+var boat = new Boat({x: canvasWidth/2, y: 25, width: 200, height: 150, velocity: 5, score: 0});
 
 function drawOcean() {
   context.drawImage(images.ocean, 0, 135);
 }
 
-var Boat = require('./boat');
-var Bullet = require('./bullet');
-var fishies = require('./generate-fish');
-var hit = require('./collision');
-// var keyEvents = require('./key-events');
-
-var bullets = [];
-
-var boat = new Boat({x: canvasWidth/2, y: 25, width: 200, height: 150, velocity: 5});
-
 canvas.addEventListener('keydown', function(key) {
-//   var keyNumber = event.keyCode;
-//   keyEvents(keyNumber);
-
   if (key.which === 37) {
     boat.moveBoatLeft(boat);
   } else if (key.which === 39) {
@@ -60,9 +51,10 @@ function filterBullets() {
   });
 }
 
-function generateFish(fishies) {
+function generateFish() {
   fishies.forEach(function (fish) {
-    fish.draw(context).move()
+    fish.draw(context)
+        .move()
         .reverseDirection(canvasWidth);
   });
 }
@@ -77,10 +69,11 @@ function generateBullets(bullets) {
 function checkForCollision() {
   fishies.forEach(function(fish) {
     bullets.forEach(function(bullet) {
-      if (hit.collision(fish, bullet)) {
+      if (collision(fish, bullet)) {
+        boat.score += fish.y;
+        console.log(boat.score)
         fishies = fishies.filter(function(eachFish){
-        return eachFish !== fish;
-        hit.updateScore();
+          return eachFish !== fish;
         });
       }
     });
@@ -88,20 +81,20 @@ function checkForCollision() {
 }
 
 function Game() {
-  clearScreen();
   startGame();
 }
 
-function startGame() {
+var startGame = function() {
+  clearScreen();
   drawOcean();
-  generateFish(fishies);
+  generateFish();
   boat.draw(context);
   generateBullets(bullets);
   checkForCollision();
-}
+};
 
 function clearScreen() {
   context.clearRect(0, 0, canvas.width, canvas.height);
-}
+};
 
 module.exports = Game;
