+++ lib/image-renderer.js
